Reworked the download page to link from "add this key to your apt configuration" to the /archive-key page, pushing the SecureApt reference link into that page. This has a streamlining affect on new users wanting to get started as it links directly to the archive-key page where there are explicit instructions for how to do the archive key verification and then the follow-up add to the apt keyring for proper archive/package verification. The way it was before, you were given the link to the more general SecureApt debian wiki page, which was a lot of information, when in most cases people just want to know what steps they need to know to do the verification/addition. However, the link to the SecureApt page is valuable to reference, it just makes more sense to be embedded in the page which explicitly deals with that.
This also means I added information about how to add the key to the
apt keyring into the archive-key page. In doing so, I removed the
dkg-specific prompt (not because I dont like it, but because it would
have been funny for me to add more commands and have to manually
construct his prompt to maintain consistency on the page). Also, added
some output of commands so people have an idea what to expect.