add keys-for-user subcommand to monkeysphere-authentication
This subcommand will output all valid key for a given user.  The
user's authorized_user_ids file will be read for OpenPGP user IDs, one
per line.  The ssh-formated RSA keys will be output to stdout.

Also included is a simple script that takes the user as it's one
argument and exec's this command.  This is something that would be
suitable for the proposed sshd AuthorizedKeysCommand.